Set up JOI validation using Middleware
